Make paths to dynamically loaded resources relative. Fixes #121 (#122)
* Make paths to dynamically loaded resources relative

When running npm run dev locally, all paths in the project start at
the root of the webserver. However, when hosting on github, the app
is loaded under a project subdirectory. This causes some resource loads to fail.

So far I just found 'service-worker.js' and some of the icons failing to
load. Since service-worker.js is just used to manage the caching API
right now, I doubt users noticed the issue, it would just cause more
network traffic.

* Update service-worker.js resources to cache

I've updated the service-worker.js paths of URLs to cache
with all of the files that get generated after running npm run build
and removed the leading slashes. That seems to be enough to get
the app to run after installing it and then turning off WiFi.

* Addedd some comments to service-worker.js to explain the intent